% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioworldr_package.R
\docType{package}
\name{BioWorldR}
\alias{BioWorldR}
\alias{BioWorldR-package}
\title{BioWorldR: A Curated Collection of Biodiversity and Species Datasets and Utilities}
\description{
{
This package provides a curated collection of biodiversity and species-related datasets (birds, plants, reptiles, turtles, mammals, bees, marine data and related biological measurements), together with small utilities to load and explore them.
}
}
\details{
BioWorldR: A Curated Collection of Biodiversity and Species Datasets and Utilities

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Curated Collection of Biodiversity and Species Datasets and Utilities
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/Monroy31039/BioWorld}
}
}
}
\author{
{
\strong{Maintainer}: Juan David Monroy \email{monroyjuandavid773@gmail.com}
}
}
