% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Extinct_mammals_tbl_df}
\alias{Extinct_mammals_tbl_df}
\title{Extinct Mammals Data}
\format{
A data frame with 85 observations and 6 variables:
\describe{
  \item{Common.name}{Common name of the species (character)}
  \item{Binomial.name}{Scientific (binomial) name of the species (character)}
  \item{Order}{Taxonomic order (character)}
  \item{Date.of.extinction}{Date or year of extinction (character)}
  \item{Former.range}{Former geographic range (character)}
  \item{Picture}{Logical flag indicating presence of a picture (logical)}
}
}
\source{
Data taken from Kaggle: \url{https://www.kaggle.com/datasets/battle11king/extinct-mammals}
}
\usage{
data(Extinct_mammals_tbl_df)
}
\description{
This dataset, Extinct_mammals_tbl_df, is a data frame containing information on 85 extinct mammal species.
It includes common and binomial names, taxonomic order, date of extinction, former geographic range,
and a flag indicating the presence of a picture. The data were read from a CSV file and no modifications have been made.
}
\details{
The dataset name has been kept as 'Extinct_mammals_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
BioWorldR package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a data frame/tibble. The original content has not been modified.
}
